requirements.txt
pandas
openpyxl
fpdf2
pdfrw
jinja2